+++ tests/testthat/test_taxaid_ORWA.R
@@ -8,27 +8,31 @@
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # 20231031, removed "DNI" if present so test passes.
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# only 1 test for dups since the same table for all 6
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# 20231206, added 6th test
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 # duplicate names ----
-test_that("trans, dups1", {
+test_that("ORWA, trans, dups1", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files.csv"))
   fn_all <- df_pickfiles$filename
   
-  i <- 1# + 6
+  ii <- 1# + 6
   
   # for (i in seq_len(length(fn_all))) {
-  fn_i <- df_pickfiles[i , "filename"]
-  df_i <- read.csv(file.path("data", "taxa_official", fn_i))
-  i_taxaid <- df_pickfiles[i, "taxaid"]
+  fn_ii <- df_pickfiles[ii , "filename"]
+  df_ii <- read.csv(file.path("data", "taxa_official", fn_ii))
+  ii_taxaid <- df_pickfiles[ii, "taxaid"]
   
   # QC
-  n_taxa_calc <- length(unique(df_i[, i_taxaid]))
-  n_taxa_qc   <- length(df_i[, i_taxaid])
+  n_taxa_calc <- length(unique(df_ii[, ii_taxaid]))
+  n_taxa_qc   <- length(df_ii[, ii_taxaid])
   
   # show mismatches
-  print(paste0("Duplicate names '", i_taxaid, "' from file '", fn_i, "'"))
-  n_taxa_table <- as.data.frame(table(df_i[, i_taxaid]))
+  print(paste0("Duplicate names '", ii_taxaid, "' from file '", fn_ii, "'"))
+  n_taxa_table <- as.data.frame(table(df_ii[, ii_taxaid]))
   n_taxa_table[n_taxa_table$Freq > 1, ]
   
   #test
@@ -36,26 +40,26 @@ test_that("trans, dups1", {
   
 })## Test ~ dup, trans
 
-test_that("attr, dups1", {
+test_that("ORWA, attr, dups1", {
   
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files.csv"))
   fn_all <- df_pickfiles$filename
   
-  i <- 1 #+ 6
+  jj <- 1 #+ 6
 
   # attribute file
-  fn_j <- df_pickfiles[i , "attributes_filename"]
-  df_j <- read.csv(file.path("data", "taxa_official", fn_j))
-  j_taxaid <- df_pickfiles[i, "attributes_taxaid"]
+  fn_jj <- df_pickfiles[jj , "attributes_filename"]
+  df_jj <- read.csv(file.path("data", "taxa_official", fn_jj))
+  jj_taxaid <- df_pickfiles[jj, "attributes_taxaid"]
   
   # QC
-  n_taxa_calc <- length(unique(df_j[, j_taxaid]))
-  n_taxa_qc   <- length(df_j[, j_taxaid])
+  n_taxa_calc <- length(unique(df_jj[, jj_taxaid]))
+  n_taxa_qc   <- length(df_jj[, jj_taxaid])
   
   # show mismatches
-  print(paste0("Duplicate names '", j_taxaid, "' from file '", fn_j, "'"))
-  n_taxa_table <- as.data.frame(table(df_j[, j_taxaid]))
+  print(paste0("Duplicate names '", jj_taxaid, "' from file '", fn_jj, "'"))
+  n_taxa_table <- as.data.frame(table(df_jj[, jj_taxaid]))
   n_taxa_table[n_taxa_table$Freq > 1, ]
   
   # test
@@ -65,7 +69,7 @@ test_that("attr, dups1", {
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # trans, taxaid ----
-test_that("trans, taxaid_1", {
+test_that("ORWA, trans, taxaid_1", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files.csv"))
   fn_all <- df_pickfiles$filename
@@ -110,7 +114,7 @@ test_that("trans, taxaid_1", {
 })## Test ~ taxaid
     
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-test_that("trans, taxaid_2", {
+test_that("ORWA, trans, taxaid_2", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files.csv"))
   fn_all <- df_pickfiles$filename
@@ -143,7 +147,7 @@ test_that("trans, taxaid_2", {
 })## Test ~ taxaid
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-test_that("trans, taxaid_3", {
+test_that("ORWA, trans, taxaid_3", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files.csv"))
   fn_all <- df_pickfiles$filename
@@ -176,7 +180,7 @@ test_that("trans, taxaid_3", {
 })## Test ~ taxaid
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-test_that("trans, taxaid_4", {
+test_that("ORWA, trans, taxaid_4", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files.csv"))
   fn_all <- df_pickfiles$filename
@@ -209,7 +213,7 @@ test_that("trans, taxaid_4", {
 })## Test ~ taxaid
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-test_that("trans, taxaid_5", {
+test_that("ORWA, trans, taxaid_5", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files.csv"))
   fn_all <- df_pickfiles$filename
@@ -242,12 +246,44 @@ test_that("trans, taxaid_5", {
 })## Test ~ taxaid
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+test_that("ORWA, trans, taxaid_6", {
+  # data files
+  df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files.csv"))
+  fn_all <- df_pickfiles$filename
+  
+  i <- 6 #+ 6
+  
+  # for (i in seq_len(length(fn_all))) {
+  fn_i <- df_pickfiles[i , "filename"]
+  df_i <- read.csv(file.path("data", "taxa_official", fn_i))
+  i_taxaid <- df_pickfiles[i, "taxaid"]
+  i_taxaid_match <- df_pickfiles[i, "calc_taxaid"]
+  
+  n_match_calc <- sum(unique(df_i[, i_taxaid_match]) %in% df_i[, i_taxaid])
+  n_match_QC <- length(unique(df_i[, i_taxaid_match]))
+  
+  # show mismatches
+  print(paste0("Unique '", i_taxaid_match, "' missing from '", i_taxaid, "'"))
+  i6 <- unique(df_i[, i_taxaid_match])[!unique(df_i[, i_taxaid_match]) %in% 
+                                         df_i[, i_taxaid]]
+  sort(i6)
+  
+  # Remove DNI
+  n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
+  
+  # test
+  testthat::expect_equivalent(n_match_calc, n_match_QC)
+  
+  # }## FOR ~ i
+  
+})## Test ~ taxaid
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # All missing taxa
-# sort(unique(c(i1, i2, i3, i4, i5)))
+# sort(unique(c(i1, i2, i3, i4, i5, i6)))
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 # attr, taxaid ----
-test_that("attr, taxaid1", {
+test_that("ORWA, attr, taxaid1", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files.csv"))
   fn_all <- df_pickfiles$filename
@@ -294,7 +330,7 @@ test_that("attr, taxaid1", {
 })## Test ~ taxaid
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-test_that("attr, taxaid_2", {
+test_that("ORWA, attr, taxaid_2", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files.csv"))
   fn_all <- df_pickfiles$filename
@@ -332,7 +368,7 @@ test_that("attr, taxaid_2", {
 })## Test ~ taxaid
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-test_that("attr, taxaid_3", {
+test_that("ORWA, attr, taxaid_3", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files.csv"))
   fn_all <- df_pickfiles$filename
@@ -370,7 +406,7 @@ test_that("attr, taxaid_3", {
 })## Test ~ taxaid
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-test_that("attr, taxaid_4", {
+test_that("ORWA, attr, taxaid_4", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files.csv"))
   fn_all <- df_pickfiles$filename
@@ -408,7 +444,7 @@ test_that("attr, taxaid_4", {
 })## Test ~ taxaid
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-test_that("attr, taxaid_5", {
+test_that("ORWA, attr, taxaid_5", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files.csv"))
   fn_all <- df_pickfiles$filename
@@ -434,6 +470,44 @@ test_that("attr, taxaid_5", {
   j5 <- unique(df_i[, i_taxaid_match])[!unique(df_i[, i_taxaid_match]) %in% 
                                          df_j[, j_taxaid]]
   sort(j5)
+  
+  # Remove DNI
+  n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
+  
+  # test
+  testthat::expect_equivalent(n_match_calc, n_match_QC)
+  
+  # }## FOR ~ i
+  
+})## Test ~ taxaid
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+test_that("ORWA, attr, taxaid_6", {
+  # data files
+  df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files.csv"))
+  fn_all <- df_pickfiles$filename
+  
+  i <- 6 # + 6
+  
+  # for (i in seq_len(length(fn_all))) {
+  fn_i <- df_pickfiles[i , "filename"]
+  df_i <- read.csv(file.path("data", "taxa_official", fn_i))
+  i_taxaid <- df_pickfiles[i, "taxaid"]
+  i_taxaid_match <- df_pickfiles[i, "calc_taxaid"]
+  
+  # attribute file
+  fn_j <- df_pickfiles[i , "attributes_filename"]
+  df_j <- read.csv(file.path("data", "taxa_official", fn_j))
+  j_taxaid <- df_pickfiles[i, "attributes_taxaid"]
+  
+  n_match_calc <- sum(unique(df_i[, i_taxaid_match]) %in% df_j[, j_taxaid])
+  n_match_QC <- length(unique(df_i[, i_taxaid_match]))
+  
+  # show mismatches
+  print(paste0("Unique '", j_taxaid, "' missing from '", i_taxaid_match, "'"))
+  j6 <- unique(df_i[, i_taxaid_match])[!unique(df_i[, i_taxaid_match]) %in% 
+                                         df_j[, j_taxaid]]
+  sort(j6)
   
   # Remove DNI
   n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
@@ -481,6 +555,6 @@ test_that("attr, taxaid_5", {
 # })## Test ~ taxaid
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # All missing taxa
-# sort(unique(c(j1, j2, j3, j4)))
+# sort(unique(c(j1, j2, j3, j4, j5, j6)))
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~