+++ tests/testthat/test_taxaid_RedLakes.R
@@ -9,25 +9,26 @@
 # 20231031, removed "DNI" if present so test passes.
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # 20231201, MODIFY for Red Lakes
+# cut to only 1 (bugs) and fish (3) [2 and 4 duplicate 1 and 3]
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 
 # duplicate names ----
-test_that("trans, dups1", {
+test_that("RedLakes, trans, dups1", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_RedLakes.csv"))
   fn_all <- df_pickfiles$filename
   
-  i <- 1# + 6
+  ii <- 1# + 4
   
   # for (i in seq_len(length(fn_all))) {
-  fn_i <- df_pickfiles[i , "filename"]
-  df_i <- read.csv(file.path("data", "taxa_official", fn_i))
-  i_taxaid <- df_pickfiles[i, "taxaid"]
+  fn_ii <- df_pickfiles[ii , "filename"]
+  df_ii <- read.csv(file.path("data", "taxa_official", fn_ii))
+  ii_taxaid <- df_pickfiles[ii, "taxaid"]
   
   # QC
-  n_taxa_calc <- length(unique(df_i[, i_taxaid]))
-  n_taxa_qc   <- length(df_i[, i_taxaid])
+  n_taxa_calc <- length(unique(df_ii[, ii_taxaid]))
+  n_taxa_qc   <- length(df_ii[, ii_taxaid])
   
   # show mismatches
   print(paste0("Duplicate names '", i_taxaid, "' from file '", fn_i, "'"))
@@ -39,26 +40,79 @@ test_that("trans, dups1", {
   
 })## Test ~ dup, trans
 
-test_that("attr, dups1", {
+test_that("RedLakes, trans, dups3", {
+  # data files
+  df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_RedLakes.csv"))
+  fn_all <- df_pickfiles$filename
+  
+  ii <- 3# + 4
+  
+  # for (i in seq_len(length(fn_all))) {
+  fn_ii <- df_pickfiles[ii , "filename"]
+  df_ii <- read.csv(file.path("data", "taxa_official", fn_ii))
+  ii_taxaid <- df_pickfiles[ii, "taxaid"]
+  
+  # QC
+  n_taxa_calc <- length(unique(df_ii[, ii_taxaid]))
+  n_taxa_qc   <- length(df_ii[, ii_taxaid])
+  
+  # show mismatches
+  print(paste0("Duplicate names '", i_taxaid, "' from file '", fn_i, "'"))
+  n_taxa_table <- as.data.frame(table(df_i[, i_taxaid]))
+  n_taxa_table[n_taxa_table$Freq > 1, ]
+  
+  #test
+  testthat::expect_equal(n_taxa_calc, n_taxa_qc)
+  
+})## Test ~ dup, trans
+
+test_that("RedLakes, attr, dups1", {
   
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_RedLakes.csv"))
   fn_all <- df_pickfiles$filename
   
-  i <- 1 #+ 6
+  jj <- 1 #+ 6
 
   # attribute file
-  fn_j <- df_pickfiles[i , "attributes_filename"]
-  df_j <- read.csv(file.path("data", "taxa_official", fn_j))
-  j_taxaid <- df_pickfiles[i, "attributes_taxaid"]
+  fn_jj <- df_pickfiles[i , "attributes_filename"]
+  df_jj <- read.csv(file.path("data", "taxa_official", fn_jj))
+  jj_taxaid <- df_pickfiles[jj, "attributes_taxaid"]
   
   # QC
-  n_taxa_calc <- length(unique(df_j[, j_taxaid]))
-  n_taxa_qc   <- length(df_j[, j_taxaid])
+  n_taxa_calc <- length(unique(df_jj[, jj_taxaid]))
+  n_taxa_qc   <- length(df_jj[, jj_taxaid])
   
   # show mismatches
-  print(paste0("Duplicate names '", j_taxaid, "' from file '", fn_j, "'"))
-  n_taxa_table <- as.data.frame(table(df_j[, j_taxaid]))
+  print(paste0("Duplicate names '", jj_taxaid, "' from file '", fn_jj, "'"))
+  n_taxa_table <- as.data.frame(table(df_jj[, jj_taxaid]))
+  n_taxa_table[n_taxa_table$Freq > 1, ]
+  
+  # test
+  testthat::expect_equivalent(n_taxa_calc, n_taxa_qc)
+  
+})## Test ~ dup, trans
+
+test_that("RedLakes, attr, dups3", {
+  
+  # data files
+  df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_RedLakes.csv"))
+  fn_all <- df_pickfiles$filename
+  
+  jj <- 3 #+ 6
+  
+  # attribute file
+  fn_jj <- df_pickfiles[i , "attributes_filename"]
+  df_jj <- read.csv(file.path("data", "taxa_official", fn_jj))
+  jj_taxaid <- df_pickfiles[jj, "attributes_taxaid"]
+  
+  # QC
+  n_taxa_calc <- length(unique(df_jj[, jj_taxaid]))
+  n_taxa_qc   <- length(df_jj[, jj_taxaid])
+  
+  # show mismatches
+  print(paste0("Duplicate names '", jj_taxaid, "' from file '", fn_jj, "'"))
+  n_taxa_table <- as.data.frame(table(df_jj[, jj_taxaid]))
   n_taxa_table[n_taxa_table$Freq > 1, ]
   
   # test
@@ -68,7 +122,7 @@ test_that("attr, dups1", {
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # trans, taxaid ----
-test_that("trans, taxaid_1", {
+test_that("RedLakes, trans, taxaid_1", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_RedLakes.csv"))
   fn_all <- df_pickfiles$filename
@@ -113,40 +167,8 @@ test_that("trans, taxaid_1", {
 })## Test ~ taxaid
     
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-test_that("trans, taxaid_2", {
-  # data files
-  df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_RedLakes.csv"))
-  fn_all <- df_pickfiles$filename
-  
-  i <- 2 #+ 6
-  
-  # for (i in seq_len(length(fn_all))) {
-  fn_i <- df_pickfiles[i , "filename"]
-  df_i <- read.csv(file.path("data", "taxa_official", fn_i))
-  i_taxaid <- df_pickfiles[i, "taxaid"]
-  i_taxaid_match <- df_pickfiles[i, "calc_taxaid"]
 
-  n_match_calc <- sum(unique(df_i[, i_taxaid_match]) %in% df_i[, i_taxaid])
-  n_match_QC <- length(unique(df_i[, i_taxaid_match]))
-  
-  # show mismatches
-  print(paste0("Unique '", i_taxaid_match, "' missing from '", i_taxaid, "'"))
-  i2 <- unique(df_i[, i_taxaid_match])[!unique(df_i[, i_taxaid_match]) %in% 
-                                         df_i[, i_taxaid]]
-  sort(i2)
-  
-  # Remove DNI
-  n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
-  
-  # test
-  testthat::expect_equivalent(n_match_calc, n_match_QC)
-  
-  # }## FOR ~ i
-  
-})## Test ~ taxaid
-
-#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-test_that("trans, taxaid_3", {
+test_that("RedLakes, trans, taxaid_3", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_RedLakes.csv"))
   fn_all <- df_pickfiles$filename
@@ -178,38 +200,7 @@ test_that("trans, taxaid_3", {
   
 })## Test ~ taxaid
 
-#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-test_that("trans, taxaid_4", {
-  # data files
-  df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_RedLakes.csv"))
-  fn_all <- df_pickfiles$filename
-  
-  i <- 4 #+ 6
-  
-  # for (i in seq_len(length(fn_all))) {
-  fn_i <- df_pickfiles[i , "filename"]
-  df_i <- read.csv(file.path("data", "taxa_official", fn_i))
-  i_taxaid <- df_pickfiles[i, "taxaid"]
-  i_taxaid_match <- df_pickfiles[i, "calc_taxaid"]
 
-  n_match_calc <- sum(unique(df_i[, i_taxaid_match]) %in% df_i[, i_taxaid])
-  n_match_QC <- length(unique(df_i[, i_taxaid_match]))
-  
-  # show mismatches
-  print(paste0("Unique '", i_taxaid_match, "' missing from '", i_taxaid, "'"))
-  i4 <- unique(df_i[, i_taxaid_match])[!unique(df_i[, i_taxaid_match]) %in% 
-                                         df_i[, i_taxaid]]
-  sort(i4)
-  
-  # Remove DNI
-  n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
-  
-  # test
-  testthat::expect_equivalent(n_match_calc, n_match_QC)
-  
-  # }## FOR ~ i
-  
-})## Test ~ taxaid
 
 
 
@@ -219,7 +210,7 @@ test_that("trans, taxaid_4", {
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 # attr, taxaid ----
-test_that("attr, taxaid1", {
+test_that("RedLakes, attr, taxaid1", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_RedLakes.csv"))
   fn_all <- df_pickfiles$filename
@@ -265,46 +256,9 @@ test_that("attr, taxaid1", {
   
 })## Test ~ taxaid
 
-#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-test_that("attr, taxaid_2", {
-  # data files
-  df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_RedLakes.csv"))
-  fn_all <- df_pickfiles$filename
-  
-  i <- 2# + 6
-  
-  # for (i in seq_len(length(fn_all))) {
-  fn_i <- df_pickfiles[i , "filename"]
-  df_i <- read.csv(file.path("data", "taxa_official", fn_i))
-  i_taxaid <- df_pickfiles[i, "taxaid"]
-  i_taxaid_match <- df_pickfiles[i, "calc_taxaid"]
-  
-  # attribute file
-  fn_j <- df_pickfiles[i , "attributes_filename"]
-  df_j <- read.csv(file.path("data", "taxa_official", fn_j))
-  j_taxaid <- df_pickfiles[i, "attributes_taxaid"]
-  
-  n_match_calc <- sum(unique(df_i[, i_taxaid_match]) %in% df_j[, j_taxaid])
-  n_match_QC <- length(unique(df_i[, i_taxaid_match]))
-  
-  # show mismatches
-  print(paste0("Unique '", j_taxaid, "' missing from '", i_taxaid_match, "'"))
-  j2 <- unique(df_i[, i_taxaid_match])[!unique(df_i[, i_taxaid_match]) %in% 
-                                         df_j[, j_taxaid]]
-  sort(j2)
-  
-  # Remove DNI
-  n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
-  
-  # test
-  testthat::expect_equivalent(n_match_calc, n_match_QC)
-  
-  # }## FOR ~ i
-  
-})## Test ~ taxaid
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-test_that("attr, taxaid_3", {
+test_that("RedLakes, attr, taxaid_3", {
   # data files
   df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_RedLakes.csv"))
   fn_all <- df_pickfiles$filename
@@ -341,45 +295,6 @@ test_that("attr, taxaid_3", {
   
 })## Test ~ taxaid
 
-#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-test_that("attr, taxaid_4", {
-  # data files
-  df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_RedLakes.csv"))
-  fn_all <- df_pickfiles$filename
-  
-  i <- 4# + 6
-  
-  # for (i in seq_len(length(fn_all))) {
-  fn_i <- df_pickfiles[i , "filename"]
-  df_i <- read.csv(file.path("data", "taxa_official", fn_i))
-  i_taxaid <- df_pickfiles[i, "taxaid"]
-  i_taxaid_match <- df_pickfiles[i, "calc_taxaid"]
-  
-  # attribute file
-  fn_j <- df_pickfiles[i , "attributes_filename"]
-  df_j <- read.csv(file.path("data", "taxa_official", fn_j))
-  j_taxaid <- df_pickfiles[i, "attributes_taxaid"]
-  
-  n_match_calc <- sum(unique(df_i[, i_taxaid_match]) %in% df_j[, j_taxaid])
-  n_match_QC <- length(unique(df_i[, i_taxaid_match]))
-  
-  # show mismatches
-  print(paste0("Unique '", j_taxaid, "' missing from '", i_taxaid_match, "'"))
-  j4 <- unique(df_i[, i_taxaid_match])[!unique(df_i[, i_taxaid_match]) %in% 
-                                         df_j[, j_taxaid]]
-  sort(j4)
-  
-  # Remove DNI
-  n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
-  
-  # test
-  testthat::expect_equivalent(n_match_calc, n_match_QC)
-  
-  # }## FOR ~ i
-  
-})## Test ~ taxaid
-
-#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~